core/remotes/docker: address fourth round of review feedback
- Reuse existing ActivityTracker from context instead of overwriting
- Close pushWriter on stall timeout to prevent resource leaks
- Move mockClock/mockTicker/NewActivityTrackerWithClock to test-only
  file (clock_test_helpers_test.go)
- Add upper bound to TestCommitStallTimeoutWithMockClock tick loop
